Removed the ability to add an upstream repository.
This is not useful unless the ability to `git fetch` is also included. This is difficult, as it requires access to the user's SSH keys. While not overly complicated when public & private keys are stored in the user's ~/.ssh directory (the requirement to enter the passphrase for the private key notwithstanding), this becomes challenging when an alternative SSH agent is used. For example, the password manager 1Password uses a custom SSH Agent that allows it to store SSH keys within its vault.

Support for alternative SSH Agents was introduced in JGit v6.0, but this version requires JDK 11.

In order to continue supporting JDK 8, we decided to omit the ability to add a remote upstream repository and to git fetch, requiring the user to perform these actions prior to invoking the application.